refactor: extract display logic to dedicated module
Phase 2.1: Create app/display module
- Extract table formatting logic from commands.rs
- Add print_task_list() for consistent task table display
- Add print_task_details() for status command output
- Add print_process_started() for run command output
- Add print_log_follow_header() for log follow mode
- Reduce commands.rs complexity by ~50 lines
- Improve single responsibility principle

All tests pass: clippy ✓, build ✓, unit tests ✓, E2E tests ✓
