+++ src/app/commands.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use crate::app::{error::Result, process, storage};
+use crate::app::{display, error::Result, process, storage};
 
 /// Run a command in the background
 pub fn run(command: Vec<String>, cwd: Option<PathBuf>, env: Vec<String>) -> Result<()> {
@@ -45,10 +45,7 @@ pub fn run(command: Vec<String>, cwd: Option<PathBuf>, env: Vec<String>) -> Resu
         &process_info.log_path,
     )?;
 
-    println!("Started background process:");
-    println!("  Task ID: {}", process_info.id);
-    println!("  PID: {}", process_info.pid);
-    println!("  Log file: {}", process_info.log_path.display());
+    display::print_process_started(&process_info.id, process_info.pid, &process_info.log_path);
 
     // Drop child to avoid zombie (we're not waiting)
     std::mem::drop(child);
@@ -60,41 +57,7 @@ pub fn run(command: Vec<String>, cwd: Option<PathBuf>, env: Vec<String>) -> Resu
 pub fn list(status_filter: Option<String>) -> Result<()> {
     let conn = storage::init_database()?;
     let tasks = storage::get_tasks_with_process_check(&conn, status_filter.as_deref())?;
-
-    if tasks.is_empty() {
-        println!("No tasks found.");
-        return Ok(());
-    }
-
-    // Print table header
-    println!(
-        "{:<8} {:<8} {:<10} {:<20} {:<30}",
-        "Task ID", "PID", "Status", "Started", "Command"
-    );
-    println!("{}", "-".repeat(80));
-
-    for task in tasks {
-        let command: Vec<String> = serde_json::from_str(&task.command).unwrap_or_default();
-        let command_str = command.join(" ");
-        let command_display = if command_str.len() > 30 {
-            format!("{}...", &command_str[..27])
-        } else {
-            command_str
-        };
-
-        let started = chrono::DateTime::from_timestamp(task.started_at, 0)
-            .map(|dt| dt.format("%Y-%m-%d %H:%M").to_string())
-            .unwrap_or_else(|| "Unknown".to_string());
-
-        println!(
-            "{:<8} {:<8} {:<10} {:<20} {:<30}",
-            &task.id[..8], // Show first 8 chars of task ID
-            task.pid,
-            task.status,
-            started,
-            command_display
-        );
-    }
+    display::print_task_list(&tasks);
 
     Ok(())
 }
@@ -111,9 +74,7 @@ pub fn log(task_id: &str, follow: bool) -> Result<()> {
 
     if follow {
         // Simple follow implementation (could be improved)
-        println!("Following logs for task {task_id} (Ctrl+C to stop):");
-        println!("Log file: {}", task.log_path);
-        println!("{}", "-".repeat(40));
+        display::print_log_follow_header(task_id, &task.log_path);
 
         // For now, just read the current content
         // A real implementation would use inotify or similar
@@ -161,31 +122,7 @@ pub fn status(task_id: &str) -> Result<()> {
 
     // This will update the status if the process is no longer running
     let task = storage::update_task_status_by_process_check(&conn, task_id)?;
-
-    println!("Task: {}", task.id);
-    println!("PID: {}", task.pid);
-    println!("Status: {}", task.status);
-
-    let command: Vec<String> = serde_json::from_str(&task.command).unwrap_or_default();
-    println!("Command: {}", command.join(" "));
-
-    let started = chrono::DateTime::from_timestamp(task.started_at, 0)
-        .map(|dt| dt.format("%Y-%m-%d %H:%M:%S").to_string())
-        .unwrap_or_else(|| "Unknown".to_string());
-    println!("Started: {started}");
-
-    if let Some(finished_at) = task.finished_at {
-        let finished = chrono::DateTime::from_timestamp(finished_at, 0)
-            .map(|dt| dt.format("%Y-%m-%d %H:%M:%S").to_string())
-            .unwrap_or_else(|| "Unknown".to_string());
-        println!("Finished: {finished}");
-    }
-
-    if let Some(exit_code) = task.exit_code {
-        println!("Exit code: {exit_code}");
-    }
-
-    println!("Log file: {}", task.log_path);
+    display::print_task_details(&task);
 
     Ok(())
 }
