+++ src/app/mod.rs
@@ -1,4 +1,5 @@
 pub mod commands;
+pub mod display;
 pub mod error;
 pub mod process;
 pub mod storage;
